use includes + rename constant file to put together all constants for environment variations
