+++ src/app/(dashboard)/layout.tsx
@@ -2,7 +2,7 @@ import ChecklistButton from '@/components/checklist/ChecklistButton'
 import withAuth, { UserSessionProps } from '@/components/hoc/withAuth'
 import Navbar from '@/components/navbar/Navbar'
 import OrganizationCard from '@/components/organizationCard/OrganizationCard'
-import { environmentsWithChecklist } from '@/constants/userCheckedSteps'
+import { environmentsWithChecklist } from '@/constants/environments'
 import { getAccountOrganizationVersions } from '@/db/account'
 import { OrganizationVersionWithOrganization } from '@/db/organization'
 import { getAllowedStudyIdByAccount } from '@/db/study'
@@ -47,15 +47,14 @@ const NavLayout = async ({ children, user: account }: Props & UserSessionProps) 
       <main className={classNames(styles.content, { [styles.withOrganizationCard]: account.organizationVersionId })}>
         {children}
       </main>
-      {accountOrganizationVersion &&
-        environmentsWithChecklist.some((env) => env === accountOrganizationVersion.environment) && (
-          <ChecklistButton
-            accountOrganizationVersion={accountOrganizationVersion}
-            clientId={clientId}
-            studyId={studyId}
-            userRole={account.role}
-          />
-        )}
+      {accountOrganizationVersion && environmentsWithChecklist.includes(accountOrganizationVersion.environment) && (
+        <ChecklistButton
+          accountOrganizationVersion={accountOrganizationVersion}
+          clientId={clientId}
+          studyId={studyId}
+          userRole={account.role}
+        />
+      )}
       <EnvironmentInitializer user={account} />
     </div>
   )
