+++ src/constants/environments.ts
@@ -1,3 +1,4 @@
 import { Environment } from '@prisma/client'
 
 export const environmentWithOnboarding: Environment[] = [Environment.BC]
+export const environmentsWithChecklist: Environment[] = [Environment.BC]
