--- src/constants/userCheckedSteps.ts
@@ -1,3 +1,0 @@
-import { Environment } from '@prisma/client'
-
-export const environmentsWithChecklist = [Environment.BC]
